Pass location to adapters for inspection
The ticket search now passes the window location and the document to
adapters for inspection (instead of only the document). This should make
it easier to detect ticket tracking systems that are only available
under a specific URL and allows extracting information from the URL
instead of the DOM.

Also: Implemented a quick and dirty version of a Trello adapter.