Add try.html for faster feedback cycle
Mimicks the boot/embed process for the extension in a regular HTML page,
passing in a list of dummy tickets so it is possible to just refresh the
page in a regular browser window to see the result.

Otherwise, we would have to go to chrome://extensions, reload switch to
a GitHub/JIRA/Trello/Pivotal page, select tickets and click the
extension icon.