Use Elm package for Octicons
Unfortunately, the JS octicons package is not easily integrated with Elm
or rather Elm's "Svg" library. At the moment, icon paths are not
exported in a way that is compatible with `Svg.Attributes.path`.

Also, if we used the original JS package, we'd have to pass the icon
data into the Elm program and hand it down to the components that need
it (possibly through multiple layers?).

The cost of using the Elm package is likely not receiving updates or not
receiving them soon after the JS package is updated.

This seems to be an acceptable trade-off at the momemt, considering the
Octicons do not seem to change a lot and us not needing any changes in
the foreseeable future.