[#106] Temporarily load Firefox addon (#132)
* [#106] Temporarily load Firefox addon

Always set an addon-id when building tickety-tick for Firefox.
It appears we should always do so since we are using `storage.sync`.

The relevant bits from MDN:

  > Extensions can explicitly set the add-on ID using the `applications`
    key in manifest.json. However, this key is usually optional.

  > Note, though, that some WebExtension APIs use the add-on ID and
    expect it to be the same from one browser session to the next. If
    you use these APIs in Firefox, then you must set the ID explicitly
    using the `applications` key.

  > Some WebExtension APIs use the add-on ID and expect it to be the
    same from one browser session to the next. If you use these APIs,
    then you must set the ID explicitly using the `applications` key.
    This applies to the following APIs: […], `storage.sync`, […]

* Use generated addon-id from Mozilla store

* Remove "firefox-local" variant

The "firefox" and "firefox-local" build variants were the same after
adjusting the manifest generation to always set the addon-id for firefox
regardless of whether we are building for a release or local development
and testing. Therefore we can just remove the distinction.