Adjust number of Jest workers on Circle CI
The default maximum number of workers in Jest is equal to the number of
cores available on a machine. However, when running inside a Circle CI
Docker container, Jest sets this value to 35 causing tests to be slowed
down.

This commit explicitly sets the maximum number of workers to the actual
number of vCPUs for the Circle CI environment (2).