Cache dependencies by `yarn.lock` checksum
It is less likely for `yarn.lock` to change because of updates not
related to dependency versions than it is for `package.json`.

Hence we should achieve better caching like this.

Also include the `arch` (architecture) in the cache key to ensure native
extensions will be compiled properly in case of container/image changes.