Fix file copying in Webpack watch mode
The `clean-webpack-plugin` is removing all files in the Webpack output
path on every rebuild in watch mode. At the same time however, the
`copy-webpack-plugin` only copies files once during the initial build
and after that only copies modified files by default.

This explicitly instructs the copy plugin to copy all files on every
rebuild so it is safe for the clean plugin to wipe the complete output
path.