[#93] Enhance commit messages (part 1)
Update default commit message format

The default commit message format now includes a commit subject line
with the ticket id and title and - if available - the ticket description
and URL in the commit message body.

The options page is updated to allow editing multi-line template
strings for commit messages, with a `textarea` that automatically
adjusts its height to its content.