Collect adapter errors & pass to root component
Ticket search now collects any adapter errors in addition to ticket
results. These are passed down to the root React component so we can
display an error message in case no tickets are found.

We assume errors are okay as long as there is at least one adapter which
returns a non-empty array of tickets.

Also, adapters can now simply return an empty array when they do not
apply in the current page context (e.g. GitHub adapter on a Jira page).