[#144] Handle adapter errors  (#150)
* Collect adapter errors & pass to root component

Ticket search now collects any adapter errors in addition to ticket
results. These are passed down to the root React component so we can
display an error message in case no tickets are found.

We assume errors are okay as long as there is at least one adapter which
returns a non-empty array of tickets.

Also, adapters can now simply return an empty array when they do not
apply in the current page context (e.g. GitHub adapter on a Jira page).

* Render error report if a problem was encountered

* Make content script errors JSON-serializable

When error objects are passed from the content script, via the
background page to the pop-up, only serializable object properties make
it to the other side.

By picking apart adapter `Error` objects, we make sure the desire error
information is available for the pop-up code.

* Make error report wording more natural