Extracted formatter into its own package
I extracted the formatter logic into its own package at
https://github.com/bitcrowd/tickety-tick-formatter

The reason is that we often have project ideas that are not
browser-plugins, but still want to format tickets into git-digestable
strings.

One example is the GitHub integration for OpenProject I am currently
working on.

A beneficial side effect is, that the formatter/template logic is
extracted and now implemented in TypeScript.