Update Jest
This updates Jest (and babel-jest) to patch a security vulnerability on
a package the old versions depended on: `set-value`.

See: https://github.com/advisories/GHSA-4jqc-8m5r-9rpr

With the update, we're now on Jest v27. On that version, Jest uses the
`node` environment as the default environment for running tests in,
instead of the slower `jsdom` environment. This means, tests which use
the DOM API need to be switched back to run in the `jsdom` environment.

See https://jestjs.io/blog/2021/05/25/jest-27#flipping-defaults for
details about the change.

For projects where only some tests require the DOM API, the Jest team
recommends to still run the whole suite in the `node` environment
(faster) and recommend the `jsdom` environment only for those tests
where it's actually needed.