Integrate LiquidJS template engine
This integrates LiquidJS for rendering branch, commit and command
templates in Tickety-Tick. This adds the following features:

- Additional filters, see: https://liquidjs.com/filters/overview.html
- Conditionals, loops etc.: https://liquidjs.com/tags/overview.html

to use inside templates.

The syntax is similar to, but not the same as the template syntax
previously used by Tickety-Tick.

When released, we should publish an announcement and migration guide,
possibly find a way of showing a notification in the browser/extension
popup to notify users about the breaking changes and the need to update
custom templates in the extension settings.