[401] Fix GitHub adapter bug label detection (#402)
GitHub's markup for issues changed, so that our detection for `bug`
labels no longer worked. This adjusts our selectors, so that we can
figure out whether an issue is labelled as bug or not.

Fixes: #401

Ticket: https://github.com/bitcrowd/tickety-tick/issues/401

On the side this also remove support for GH projects.

Projects work differently in GitHub now. They belong to the
user/organization, not to one repository. We need to handle them
differently. The current adapter's implementation does not work any
more.