Refactor background script
Instead of relying on the background page instance (which is not
recommended, see link to the docs), we communicate with the background
script (or service worker script in Chrome) via message passing.

This way, we can use the same script in both browsers.

Mozilla docs on the topic: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Background_scripts#update_calls_for_background_script_functions), we