[#443] GitHub adapter broken after GitHub redesign
This updates the GitHub adapter to work with the redesigned issue pages.
As we do not know now GitHub Enterprise looks like for the majority of
users, we keep the previous logic for the "legacy" strategy of the
adapter.

To keep the change and the adapter simple, I dropped support selecting
issues on the issues page. Instead, we only support the issue show pages
now. On the long run, it may make sense to look into supporting GitHub
projects too.

The new GitHub adapter supports GitHub issue types and won't look for
the "bug" label any more.

Ticket: https://github.com/bitcrowd/tickety-tick/issues/443

Fixes: #443